[QT-352] Always output exec information (#76)
When executing commands that output information we should always display
it in the operation summary. This change ensures the `output` and `exec`
scenario sub-commands are always displayed in the operation summary in
the human readable mode.

* Always display the output of exec sub-commands in the operation
  status.
* Always configure the exec operation runner with a reference to the
  scenario Terraform module.
* Bump version
* Use the diagnostics helper function to determine the status for all
  scenario sub-command operations.
* Update acceptance test runners to use different ports to allow for
  better parallel acceptance test execution.

+++ internal/operation/operator_local.go
@@ -117,8 +117,10 @@ func (o *LocalOperator) Stop() error {
 	defer o.mu.Unlock()
 	o.running = false     // Prevent new work requests from being dispatched
 	o.drainWorkReqQueue() // Drain any queued operations
-	o.ctxCancel()         // Cancel in-flight operations, kill operation workers, drain the event queue
-	o.publisher.Stop()    // Turn off event publisher
+	if o.ctxCancel != nil {
+		o.ctxCancel() // Cancel in-flight operations, kill operation workers, drain the event queue
+	}
+	o.publisher.Stop() // Turn off event publisher
 	return nil
 }
